// src/app/models/add-post.model.ts
export interface AddPostRequest {
    title: string | undefined;
    content: string | undefined;
    summary: string | undefined;
    urlHandle: string | undefined;
    featuredImageUrl: string | undefined;
    visible: boolean | undefined;
    author: string | undefined;
    publishDate: string | undefined;
    updatedDate: string | undefined;
}